Added back in Colliders and Rigid Body logic for wall collisions
Hopefully this works better this time! I think the sphere collider on the characters will be good.  Also upped the mass to avoid being pushed.
But note that the sphere is so big (to pick up triggers) that it just barely doesn't fit through the door, so when you try to go through, it sort of pushes you through to the other side and the character floats a little